# examples/blog/internal/transport/http/auth.go
// Package http provides HTTP handlers for the service.
// ============================================================================
// AUTO-GENERATED by ANG (Architectural Normalized Generator)
// Version: 0.1.127
// InputHash: 2280e6dff5bcac137d5e4ca79079af522793ed3c7f1301a7955910011cd5b7e5
// CompilerHash: 294b50c17c5868adf98ff6e7e34d9b0ddcf67156d863f2f508ad57df44889164
// DO NOT EDIT - changes will be overwritten on next generation
// GENERATED: true
// ============================================================================

package http

import (
	"crypto/sha256"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/example/blog/internal/pkg/errors"
	"github.com/example/blog/internal/port"
	"github.com/go-chi/chi/v5"
)

func RegisterAuthRoutes(r chi.Router, svc port.Auth) {
	r.With(MaxBodySizeMiddleware(1048576)).Post("/auth/login", handlerLogin(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware).Get("/auth/profile", handlerGetProfile(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware).Put("/auth/profile", handlerUpdateProfile(svc))
	r.With(MaxBodySizeMiddleware(1048576)).Post("/auth/register", handlerRegister(svc))
}

// handlerLogin
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerLogin(svc port.Auth) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.LoginRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.Login(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerGetProfile
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerGetProfile(svc port.Auth) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.GetProfileRequest
		if v := r.URL.Query().Get("userid"); v != "" {
			req.UserID = v
		}
		if req.UserID == "" {
			req.UserID = CurrentUserID(r)
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.GetProfile(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		var out any
		out = resp
		data, err := json.Marshal(out)
		if err != nil {
			errors.WriteError(w, r, errors.New(http.StatusInternalServerError, "Internal Server Error", "Failed to encode response"))
			return
		}
		etag := fmt.Sprintf("W/\"%x\"", sha256.Sum256(data))
		w.Header().Set("ETag", etag)
		if match := r.Header.Get("If-None-Match"); match == etag {
			w.WriteHeader(http.StatusNotModified)
			return
		}
		_, _ = w.Write(data)

	}
}

// handlerUpdateProfile
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerUpdateProfile(svc port.Auth) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.UpdateProfileRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if req.UserID == "" {
			req.UserID = CurrentUserID(r)
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.UpdateProfile(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerRegister
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerRegister(svc port.Auth) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.RegisterRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.Register(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}